Rework some stuff
