+++ sentry-ruby/lib/sentry-ruby.rb
@@ -22,7 +22,6 @@ require "sentry/hub"
 require "sentry/background_worker"
 require "sentry/session_flusher"
 require "sentry/cron/monitor_check_ins"
-require "sentry/spotlight"
 
 [
   "sentry/rake",
